requirements.txt
fastapi
uvicorn[standard]
pydantic
paho-mqtt
python-dotenv
pyserial
evdev
